add gtest/ctest and ci configs (.travis.yml and appveyor.yml)
* add shared drishti-upload  hunter config
* add automated ctest/gtest w/ optional gpu (ogles_gpgpu) support
* add CI configurations: .travis.yml and appveyor.yml
