+++ src/CMakeLists.txt
@@ -48,6 +48,17 @@ hunter_add_package(sse2neon)
 find_package(sse2neon CONFIG REQUIRED)
 list(APPEND ACF_SDK_3RDPARTY_PKG_LIBS sse2neon::sse2neon) # Add library
 
+# We use aglet for an "off screen" OpenGL context.
+# This currently provides 3 options:
+#   * GLFWContext for most desktop systems (can provide Window)
+#   * GLContextIOS
+#   * GLContextAndroid
+# Note: Android emulator supports GPU use
+if(ACF_DO_GPU_TESTING AND ACF_BUILD_OGLES_GPGPU)
+  hunter_add_package(aglet)
+  find_package(aglet CONFIG REQUIRED)
+endif()
+
 hunter_add_package(Sugar) 
 
 add_subdirectory(lib)
@@ -56,5 +67,5 @@ add_subdirectory(lib)
 ## Applications ##
 ##################
 
-add_subdirectory(app)
 
+add_subdirectory(app)
