+++ src/lib/acf/ut/CMakeLists.txt
@@ -1,21 +1,18 @@
 # Copyright (c) 2015-2017, Ruslan Baratov, David Hirvonen
 # All rights reserved.
 
-set(test_name DrishtiACFTest)
-set(test_app test-drishti-acf)
+set(test_name ACFTest)
+set(test_app test-acf)
 
-set(drishti_acf_ut_srcs test-Detector.cpp test-drishti-acf.cpp)
+set(acf_ut_srcs test-acf.cpp)
 
-add_executable(${test_app} ${drishti_acf_ut_srcs})
+add_executable(${test_app} ${acf_ut_srcs})
 set_property(TARGET ${test_app} PROPERTY FOLDER "app/tests")
-target_link_libraries(${test_app} PUBLIC
-  ${OpenCV_LIBS}
-  GTest::gtest
-  drishtisdk)
+target_link_libraries(${test_app} PUBLIC acf GTest::gtest ${OpenCV_LIBS})
 
-if ((${DRISHTI_DO_GPU_TESTING}) AND DRISHTI_BUILD_OGLES_GPGPU)
+if ((${ACF_DO_GPU_TESTING}) AND ACF_BUILD_OGLES_GPGPU)
   target_link_libraries(${test_app} PUBLIC aglet::aglet)
-  target_compile_definitions(${test_app} PUBLIC DRISHTI_ACF_DO_GPU=1)
+  target_compile_definitions(${test_app} PUBLIC ACF_DO_GPU=1)
 endif()
 
 # GTest + CTest
