--- src/lib/acf/ut/test-drishti-acf.sh
@@ -1,10 +1,0 @@
-#!/bin/bash
-
-# TODO
-ARGUMENTS=(
-    "${DRISHTISDK}/assets/images/lena512color.png"
-    "${DRISHTISDK}/assets/images/lena512gray.png"
-    "${DRISHTISDK}/assets/drishti_face_inner_48x48.mat"
-)
-
-echo ${ARGUMENTS[@]}
